+++ client/src/helper/urls.js
@@ -58,14 +58,17 @@ const urls = {
   mailboxManage: () => '/?page=mailbox&a=manage',
   mailboxMailto: (email) => `/?page=mailbox&mailto=${email}`,
   mailboxOldAttachment: (emailId, attachmentIndex) => `/?page=mailbox&a=dlattach&mid=${emailId}&i=${attachmentIndex}`,
-  map: ({ storeId = null, foodSharePointId = null } = {}) => {
-    let path = '/karte'
+  map: ({ storeId = null, foodSharePointId = null, markers = null } = {}) => {
+    const params = []
     if (storeId) {
-      path += `?bid=${storeId}`
+      params.push(`bid=${storeId}`)
     } else if (foodSharePointId) {
-      path += `?fspId=${foodSharePointId}`
+      params.push(`fspId=${foodSharePointId}`)
     }
-    return path
+    if (markers?.length > 0) {
+      params.push(`loadMarkers=${markers}`)
+    }
+    return '/karte' + (params.length > 0 ? '?' + params.join('&') : '')
   },
   newsFromIT: () => 'https://foodsharing.freshdesk.com/support/solutions/folders/77000160479',
   vision: () => '/ueber-uns',