+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Reduce complexity of RegionTransaction.GetJoinMessage() by session information !3242 @KoeMai
 - Remove unneeded u_loadCoords function !3277 @stcz
 - Modernized EventControl to a Symfony controller !3260 @chriswalg
+- Refactor the nearby baskets list to vue !3294 @alex.simm
 
 ## Dev/Test/CI stuff
 - Added backend support for getting, adding and updating achievements. #1724 !3317 @AntonBallmaier